Refactor code based on reviewer feedback
